requirements.txt
# Google Sheets
gspread
google-auth
google-auth-oauthlib

# HTTP
requests

# Date handling
python-dateutil
